Deletion feature improved:
If an article with less than 500 characters is to be deleted, at most the
first 150 characters are automatically suggested as a deletion reason. This
should automate the frequent practice of pasting the text of nonsense
articles into the deletion reason.

If an article with a history is to be deleted, a warning is inserted,
linking to the article history (so as to avoid accidentally deleting a
vandalized page with a valid history).

If a blanked article with a history is to be deleted, and the next earlier
revision contains less than 500 characters, at most the first 150 characters
are automatically suggested as a deletion reason. Nonsense pages are often
blanked, just checking the current revision for nonsense is not good enough.

New texts in Language.php for this feature:
- excontent, exblank, exbeforeblank, historywarning


git-svn-id: file:///home/git-svn-mirror/pushmi-svn-mirror/trunk@1347 dd0e9695-b195-4be7-bd10-2dea1a65a6b6