Added a user_real_name column to the user table, and added a patch SQL file to add this field.
User.php stores and fetches the field, and has accessors for it.

User login allows setting the field on account creation.

The Preferences page allows changing the real name.

The labels are available for the real name, and the explanation of the email
field on login has been expanded to include an explanation of the real name
field, too.

Update script checks for the field, and adds it if it's missing.


git-svn-id: file:///home/git-svn-mirror/pushmi-svn-mirror/trunk@3204 dd0e9695-b195-4be7-bd10-2dea1a65a6b6