Fix for [ 988793 ] Pages in category saved as links when category page edited
Disable link cache while running the categoryMagic list to make sure that
pages under the category are not inadvertently recorded in the link tables
when the category page is edited.

This already seems to be working in head without this, but I can't find the
magic bit that does it.


git-svn-id: file:///home/git-svn-mirror/pushmi-svn-mirror/branches/REL1_3@4528 dd0e9695-b195-4be7-bd10-2dea1a65a6b6