The last checkin changed 'copyrightwarning' in a way that broke if the wiki wasn't configured to provide copyright metadata, and requires editing or rebuilding the message if the configuration is changed.
Now uses the value in existence at runtime so it can be reconfigured more
easily, and adds a 'copyrightwarning2' message which changes the wording
a bit to be used when no copyright metadata is provided.

This change should be more upwards-compatible. Also, changed 'copyrightwarning'
to be wikitext instead of raw HTML; this shouldn't break any of the built-in
texts but might requires adjustment to customized versions.


git-svn-id: file:///home/git-svn-mirror/pushmi-svn-mirror/trunk@4842 dd0e9695-b195-4be7-bd10-2dea1a65a6b6