ExternalEditor helper reference implementation (pre-alpha)
This helper application will fetch a resource such as a wiki page or an
image, load it into an application like a text editor or an image processor,
and save it back to the server. It can be associated with the MIME type
application/external-editor to be directly used by supporting software
like MediaWiki.


git-svn-id: file:///home/git-svn-mirror/pushmi-svn-mirror/trunk/phase3@7858 dd0e9695-b195-4be7-bd10-2dea1a65a6b6