* (bug 2018) Fix deletion for new schema, make work on MySQL 3 again
The archive table now has an ar_text_id field which points at the
text record containing the deleted revision's text. Older archive
records containing self-contained text are still supported and
will be restored by adding a new revision.

For now, revision and page records are still removed on deletion,
but text records are left intact. This will keep block compression
and immutable alternate storage backends working relatively cleanly.
A rev_deleted flag field is reserved in revision for further future
changes but that won't happen in the 1.5 timeframe.

There is no longer a delete-on-join which was present in earlier
1.5 revisions, so deletion should work on MySQL 3.x again.


git-svn-id: file:///home/git-svn-mirror/pushmi-svn-mirror/trunk@8772 dd0e9695-b195-4be7-bd10-2dea1a65a6b6