Initial proof of concept functionality for version tagging/snapshotting Snapshots store versions of linked resources at snapshot time in a table and force reading of those versions for embedded templates when the snapshot is displayed.
For the moment snapshots can be manually defined at Special:Snapshot,
and viewed using ?tag=review or whatever while viewing. Work yet to go
on interfacing with parser cache, logging, review, clean systems, etc.

Images aren't snapshotted currently because it's difficult to refer to
image versions usefully at the moment. When the new system goes in this
should be handlable by storing the hash key of the file much as the
page revision numbers are stored now.



git-svn-id: file:///home/git-svn-mirror/pushmi-svn-mirror/branches/snapshot-work@15516 dd0e9695-b195-4be7-bd10-2dea1a65a6b6