* Add title prefix search for Special:Undelete * Remove full-archive list from Special:Undelete * Undelete page list can use plural marker
Integrating the log more cleanly might be nice.
I did a little experimenting with joining on logging -> archive, but found
that it wanted to do a big expensive filesort. Additionally we may want
to include the filearchive table in here to turn up image versions in
the archive that don't have text revisions attached.

Prefix search has a hardcoded limit of 100, no paging.


git-svn-id: file:///home/git-svn-mirror/pushmi-svn-mirror/trunk/phase3@19365 dd0e9695-b195-4be7-bd10-2dea1a65a6b6