* (bug 10971) `aufrom` parameter doesn't work with spaces * Fix username handling issue with `auprefix` parameter
Usernames are *not* stored in "key" format, and treating them as such will cause various comparisons to misbehave. This was preventing `aufrom` from including the username (as it was supposed to) and would have broken `auprefix` for all comparisons containing spaces.

As a side-effect, both these parameters will now treat underscores as equivalent to spaces.

git-svn-id: file:///home/git-svn-mirror/pushmi-svn-mirror/trunk/phase3@24870 dd0e9695-b195-4be7-bd10-2dea1a65a6b6