I am starting to get kinda pissed off with the comment spam on my wiki. It's time to fight back!
At the moment we have confirmEdit, which is great for stopping links being automatically inserted, and stopping spammers from automatically creating accounts, but turning it on for every edit it bit too high a barrier. And we have $wgEnableSorbs, which is great for stopping open proxies.

But there is a new category of wiki comment spam emerging, one that adds random gibberish to your wiki, without any web links, and that is often not a known open proxy. It adds words like "laletoalca" and "rolerr" and "notac4tc" and "virochiorboa" and "oloreldell" and "c4teltlir". (What the fuck does that even mean, and why would anyone bother writing a bot to add it?!) Maybe the idea is to then add links later - I don't know, but it is annoying, and from my perspective the incidence of this seems to be steadily escalating.

So this extension takes it one level further, whilst still allowing anon edits, as it asks the Project HoneyPot DNSBL whether an IP address corresponds to a known "comment spammer" (which is a new category they seem to have added recently). The reason I selected honeypot rather than another black list is that for the last batch of spam comments that got through on my wiki, they were all listed as "comment spammers" on honeypot, and usually within the first 10 Google hits for those IP addresses. (i.e. based purely on past performance).

Extension considered totally experimental. Use at own risk. Requires a honeypot key to use. Steps to get this:
* Create an account - http://www.projecthoneypot.org/create_account.php
* Verify your account by visiting the emailed URL.
* Create a key - I think the URL is: http://www.projecthoneypot.org/httpbl_configure.php
* Then add to your LocalSettings.php, add:
<code>
$wgHoneyPotKey = 'inuvziahgprwh';       // (or whatever your key is that you got in the step above)
$wgCommentSpammerLog['allowed'] = true; // only if you want logs of allowed edits in [[Special:Log]], otherwise omit.
$wgCommentSpammerLog['denied']  = true; // only if you want logs of denied  edits in [[Special:Log]], otherwise omit.
require_once( "{$IP}/extensions/CommentSpammer/CommentSpammer.php" );
</code>

Any patches or updates welcome - and if you have commit access, you don't need to ask me (in fact, I'd really prefer it if you didn't) - just modify away, and please add your name as an author.
Also if you're looking for something to work on, some possible ideas for improvement include:
* Reporting back to projecthoneypot.org when a known spammer re-offends (to keep their block alive).
* Reporting back to projecthoneypot.org when someone is blocked for spamming (to make it harder for them to attack other wikis).
* Automatically reviewing anon edits 24 or 48 hours after they are added, and auto-reverting those from now-known comment spammers (this stops the problem where someone has just started spamming, and they attack your wiki first - this way you get retroactive spam protection too, from the people they spam after you).
* <Insert your cool idea here>! 

(In short, you want to allow people to fight spam collectively, rather than individually, or to find other ways to reduce the cost of being spammed, but increase the cost to the spammer)


git-svn-id: file:///home/git-svn-mirror/pushmi-svn-mirror/trunk@27095 dd0e9695-b195-4be7-bd10-2dea1a65a6b6