Allow user to attach to an existing account at login
We connect an OpenID with a user account in the wiki. Previously, it
had to be a _new_ account. Now, after the login is finished, we let
you choose an account to connect to.

Note that if the OpenID nickname is set, and available, it will be
used automatically, without the user being asked.



git-svn-id: file:///home/git-svn-mirror/pushmi-svn-mirror/trunk@31875 dd0e9695-b195-4be7-bd10-2dea1a65a6b6