Revert 33072 for now -- changes to ucuserprefix use
Using the user_timestamp index means you can't do an indexed search on the name prefix...
Basically the same as scanning raw on rev_timestamp, but with logged-in edits skipped so you don't scan *quite* as huge an amount of the table.
Since there's no index for the username prefix, the more specific the prefix, the *more* rows you'll have to scan through to fill out the result set, so the worse the performance will get.



git-svn-id: file:///home/git-svn-mirror/pushmi-svn-mirror/trunk@33079 dd0e9695-b195-4be7-bd10-2dea1a65a6b6