* (bug 13725) Upload form watch checkbox state set correctly with wpDestFile
r33330 just disabled the check for wpDestFile, so that the 'watch pages i create'
setting would watch *all* uploads. The system now checks the actual value, so
it sets the watch checkbox if the requested file doesn't exist locally and you
have it set for creations, or if you've already got the file watched.

There are still some minor issues:
* Since you have the chance to change the target filename on the form, the
  initial checkbox state can get out of sync if you switch from an empty file
  to an existing file or vice-versa.
* Unlike the edit form, the upload form doesn't _unwatch_ a previously-watched
  page when the box is unchecked.




git-svn-id: file:///home/git-svn-mirror/pushmi-svn-mirror/trunk@33387 dd0e9695-b195-4be7-bd10-2dea1a65a6b6