Revert r36093 (bug 14404).
Regressions caused by this:
"Use an anonymous user to get the parser options and the parser cache key, thus wikitext rendering is not depending of some sepcific user options such as lang, ..."

User-specific options such as stub threshold were still applying in the parser, but not taken into account in the parser hash key. As a result, the caches were corrupt, saving different options into the anonymous-default options cache.



git-svn-id: file:///home/git-svn-mirror/pushmi-svn-mirror/trunk@36185 dd0e9695-b195-4be7-bd10-2dea1a65a6b6