API: Added bkip parameter to list=blocks. This more than fixes bug 14405.
About the database query: I know it filesorts the result set, but that shouldn't cause any trouble as the result set is always small. To be more specific: it cannot be larger than 17 rows. This is because bkip=foo only grabs (range) blocks that contain foo *in its entirety*. Therefore, in the worst case foo is a single IP and the result set will consist of a single block, a /31 range block, a /30, etc. to /16 (blocks that cross /16 , which adds up to 17 blocks (note that there's only one /N range that a /N+K range is part of).

git-svn-id: file:///home/git-svn-mirror/pushmi-svn-mirror/trunk/phase3@36452 dd0e9695-b195-4be7-bd10-2dea1a65a6b6