* (bug 14646) Fix some double-escaping of HTML in feed output
Log entry action text was being double-escaped by running HTML through the comment parser.
Log action text is now output as straight HTML.

The link for long diffs was also coming out wrong on English Wikipedia due to overloading of messages.
Switched from 'differenes' (which is a label text which had gotten turned into an HTML link there) to 'showdiff', which is a plaintext button label. (Might consider making it a separate message which is a little more descriptive.)


git-svn-id: file:///home/git-svn-mirror/pushmi-svn-mirror/trunk@37202 dd0e9695-b195-4be7-bd10-2dea1a65a6b6