* add i18n file, removed direct usage of $wgMessageCache * added descriptionmsg * bumped version * removed trailing whitespace * add support for CategoryWatch in Translate
i18n issue: 'categorywatch-emailbody' is a lego style message. For localisation in many languages other than English this contruction does not allow for proper localisation. Best drop 'This message is to notify you that at $3 user $4 $5.' and incorporate it into 'categorywatch-cat*' leading to something like "This message is to notify you that at $time user $username removed $objectlink from category $catname" for 'categorywatch-catsub'.

git-svn-id: file:///home/git-svn-mirror/pushmi-svn-mirror/trunk@40404 dd0e9695-b195-4be7-bd10-2dea1a65a6b6