Prohibit empty page titles at a low level
This adds a sanity check to EditPage::doEdit() that throws an exception if the Title's name (sans namespace) is empty.  Apparently the API edit module doesn't handle this error correctly at a high level, as evidenced by page 19405691 on enwiki.  I didn't try to test whether this extra check stops the particular error, but it doesn't hurt in any case.


git-svn-id: file:///home/git-svn-mirror/pushmi-svn-mirror/trunk/phase3@41085 dd0e9695-b195-4be7-bd10-2dea1a65a6b6