Revert r41364 -- broke 22 parser test cases with change of alt behavior.
The caption was originally defined *as* the alt text (defaulting to the image file name if there is no alt text). Note that a separate caption text is only displayed in some display modes ('frame' and 'thumb', iirc), and not by default.

Please run the parser tests and check the effect you have on them. If it's really an appropriate change, then update the test cases. If you're not sure, consider backing out pending further discussion. :)

It might be appropriate to not set the 'alt' attribute for frame/thumb cases, but definitely not for inline images where we already have a way of setting the alt text which you're removing!



git-svn-id: file:///home/git-svn-mirror/pushmi-svn-mirror/trunk/phase3@41407 dd0e9695-b195-4be7-bd10-2dea1a65a6b6