Backing out r42575 "Add quick script to update to the latest SVN version Wikimedia is using."
This will update backwards as well as forwards, which is a little scary. :)

Since it could be pretty surprising and make your life harder to recover from if you run it by accident (losing mixed versions, encountering new update conflicts, etc), I'd recommend backing it out for now. Also 'scap' is a term of art specific to our deployment system, and is not used appropriately here. ;) 



git-svn-id: file:///home/git-svn-mirror/pushmi-svn-mirror/trunk/phase3@42591 dd0e9695-b195-4be7-bd10-2dea1a65a6b6