Create new, better tooltip/accesskey methods
Old Linker::tooltipAndAccesskey() and Linker::tooltip(), which pass
around raw bits of HTML, are now wrappers for Linker::titleAttrib() and
Linker::accesskey().  Actual usages have not been changed for now.

Note that this might have some slight performance implication, since
the accesskey message will typically be retrieved twice per view
instead of once: once for the accesskey attribute, once for the bit
we're appending to the title attribute.  This used to be avoided by
using a single tooltipAndAccesskey() method, but that's pretty ugly.
Of course, the second hit should reach the local accesskey cache and
not result in a second memcached hit, but I recall Tim saying that that
still takes 100 us or something unreasonable like that.

Causes no new parser test failures, and I didn't notice any difference
in the UI with a quick check.


git-svn-id: file:///home/git-svn-mirror/pushmi-svn-mirror/trunk/phase3@42702 dd0e9695-b195-4be7-bd10-2dea1a65a6b6