* (bug 2585) HTTP 404 return code is now given for a page view if the page   does not exist, allowing spiders and link checkers to detect broken links.
This is less expansive than the old 2005 implementation (r11307), hitting
only page views (won't affect action=edit) and doesn't attempt to cover
error conditions either (many of which should probably return a different code).

Pages which exist in the DB or return true for Title::isAlwaysKnown() such
as file pages for existing files, as well as category pages that exist, are
treated as existing by returning true for Article::hasViewableContent().



git-svn-id: file:///home/git-svn-mirror/pushmi-svn-mirror/trunk@44919 dd0e9695-b195-4be7-bd10-2dea1a65a6b6