Partial revert of r41018 "Wrap $log->addEntry() in transaction"
In addition to wrapping the given function call in a transaction, the
commit also (without explanation or apparent purpose) moved a block of
code to a totally different place, which broke it entirely.
Specifically, it moved the update of the category table on article
delete to AFTER THE CATEGORYLINKS ROWS WERE ALREADY DELETED, which meant
that it found the article was in no categories, and article deletions
silently failed to decrement the category count (bug 17155).

Could you PLEASE not commit all sorts of stuff under the guise of
one-line commit messages that don't actually relate to what you did.
Thank you.

git-svn-id: file:///home/git-svn-mirror/pushmi-svn-mirror/trunk/phase3@47319 dd0e9695-b195-4be7-bd10-2dea1a65a6b6