(bug 16950) Show move log when recreating a page
Showing the move log as well as delete log helps with pages that were
moved with no redirect, or moved and then the redirect deleted with a
confusing message.  Previously, such pages would appear to have never
existed to a user trying to create them.

Patch by church of emacs, with trivial modifications by me.

git-svn-id: file:///home/git-svn-mirror/pushmi-svn-mirror/trunk@49821 dd0e9695-b195-4be7-bd10-2dea1a65a6b6