Cleanup in MagicWord::$mVariableIDs, removed IDs that aren't handled in Parser::getVariableValue().
Side effect: Removed "Pages in namespace (Magic word disabled currently)" parser test added in r17687 by nickj (was passing; I know). The problem is that the expected result is an empty string, but since it's disabled by default with $wgAllowSlowParserFunctions = false (i.e. not added to parser's list of magic words IDs) it should return a red link to [[Template:PAGESINNAMESPACE:]] (the empty string is produced because "{{PAGESINNAMESPACE:}}" is handled as variable and not as parser function).


git-svn-id: file:///home/git-svn-mirror/pushmi-svn-mirror/trunk@50157 dd0e9695-b195-4be7-bd10-2dea1a65a6b6