Add --setversion option to parserTests
This overrides the version string stored in the database with --record.
When using git-svn, you can do

    php maintenance/parserTests.php --quick --quiet --record \
	--setversion=`git rev-parse HEAD`

to have comprehensible version info stored for the purposes of later
bisection.  Previously just "1.16alpha" would be stored, which is
useless for tracking down the old revision -- you'd have to try guessing
based on date.

It might be useful for SpecialVersion::getVersion() to support git-svn,
but that's a more complicated and intrusive change for an admittedly
small minority, so I avoided it for now.

git-svn-id: file:///home/git-svn-mirror/pushmi-svn-mirror/trunk@50640 dd0e9695-b195-4be7-bd10-2dea1a65a6b6