Remove KHTMLFixes.css
This was added more than five years ago, in r3532.  It was being loaded
for WebKit as well as KHTML, and I could notice no difference in either
Chrome or Konqueror whether it was loaded or not.  It was responsible
for causing display errors with an HTML 5 doctype in WebKit.

git-svn-id: file:///home/git-svn-mirror/pushmi-svn-mirror/trunk/phase3@53141 dd0e9695-b195-4be7-bd10-2dea1a65a6b6