Re-enable $wgHtml5
The rendering issue that prompted r53137 has been fixed in r53141.
However, everyone should be on the lookout for browsers becoming
slightly more standards-compliant when this is enabled, possibly
breaking some things.  Since there are no known issues, it should be
safe to re-enable so it can get further testing.  See this page for info
on the effects of different doctypes:

http://hsivonen.iki.fi/doctype/

git-svn-id: file:///home/git-svn-mirror/pushmi-svn-mirror/trunk@53142 dd0e9695-b195-4be7-bd10-2dea1a65a6b6