Add support for using bash's built-in readline or equivalent for single-line input editing in eval.php etc, if the PHP readline extension is not available.
This is handy when running on Mac OS X with Apple's default PHP installation...
If bash cannot be found, will fall back to the fgets() we did before -- a horrible realm in which arrow keys and backspace don't work.

Note that this isn't quite as good as full readline support, since we don't get to retain entry history.


git-svn-id: file:///home/git-svn-mirror/pushmi-svn-mirror/trunk/phase3@54422 dd0e9695-b195-4be7-bd10-2dea1a65a6b6