Revert r54585 "Fix bug 19261 by saving the username as the signature if it is submitted empty"
Looks like the base problem is that empty is being interpreted as "put empty in the sig" instead of as "use default sig" at signature replacement time.
We shouldn't be saving the username into preferences if it hasn't been explicitly typed; if the user changes their name, they should automatically pick up the new value.


git-svn-id: file:///home/git-svn-mirror/pushmi-svn-mirror/trunk/phase3@55306 dd0e9695-b195-4be7-bd10-2dea1a65a6b6