EditToolbar: (bug 20715) Search&Replace messes up in Opera
This was caused by Opera treating newlines as one character in $textarea.val().indexOf() , but as two characters in $textarea.setSelection(). Worked around by an ugly temporary hack that detects whether we're on a broken browser and replacing \n with \r\n in $textarea.val() . Will redo properly after discussing with Trevor.

git-svn-id: file:///home/git-svn-mirror/pushmi-svn-mirror/trunk@56723 dd0e9695-b195-4be7-bd10-2dea1a65a6b6