Specify percentage width on Modern search box
As discussed on IRC.  This reverts r62794, and is intended to be a
better fix for bug 19194.  By removing the display: block, we work
around the WebKit bug <https://bugs.webkit.org/show_bug.cgi?id=27982>.
But then on some font sizes, the text box becomes too narrow, so the
"Go" button winds up on the same line as the search box.  thedj reported
this as a problem both with r62794 and with simple removal of display:
block, but said it was fine with width: 85%, which also looks fine to
me.

git-svn-id: file:///home/git-svn-mirror/pushmi-svn-mirror/trunk@62795 dd0e9695-b195-4be7-bd10-2dea1a65a6b6