Bark with something more than "Failed" if !$user or if we get a PasswordError
The root cause of this issue is failed validation earlier in the
installer. It's still a good idea to bark with specific internal
errors at this point, in case something bad happens that we hadn't
thought to check for.

In my case I got "Failed" because $user === $password. I couldn't
produce a failure in User::newFromName() but sanity checks are good.

git-svn-id: file:///home/git-svn-mirror/pushmi-svn-mirror/branches/new-installer@65698 dd0e9695-b195-4be7-bd10-2dea1a65a6b6