Don't allow boolean HTML5 attribs into XHTML1
The code didn't handle the case where a non-associative array was passed
for the attributes, like array( 'autofocus' ) instead of array(
'autofocus' => '' ).  In retrospect, allowing this syntax was a bad
decision and I wish I hadn't.  Associative arrays shouldn't pretend to
be lists.  Probably too much trouble to change it now.

git-svn-id: file:///home/git-svn-mirror/pushmi-svn-mirror/trunk@67869 dd0e9695-b195-4be7-bd10-2dea1a65a6b6