Recognize webm and matroska files. See also Bug 23888
Currently all webm files are stored as video/webm. It is not possible to detect 
wether this file is an audio file without using a full parser. This is why We should 
really move mime and mediatype accessors to the MediaHandlers.

Using video/x-matroska for MKV files. There is no official mime for MKV (though the 
webm isn't official either, but everyone is already using it apparently).



git-svn-id: file:///home/git-svn-mirror/pushmi-svn-mirror/trunk/phase3@70103 dd0e9695-b195-4be7-bd10-2dea1a65a6b6