Improve $wgExperimentalHtmlIds support
Now works (again?) in IE6.  That didn't like getting non-legacy ID's on
redirects -- it doesn't interpret them as UTF-8, so anything non-ASCII
breaks.  If this works in all the browsers I test in, I'll enable it --
it will produce *much* nicer-looking anchors for non-English wikis, and
also section names with English punctuation.

git-svn-id: file:///home/git-svn-mirror/pushmi-svn-mirror/trunk@70524 dd0e9695-b195-4be7-bd10-2dea1a65a6b6